chore: move constants effectifs
